LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/static-router-race-network-and-fetch-handler.https-expected.txt


PASS Main resource load matched the rule with race-network-and-fetch-handler source, and the fetch handler response is faster than the server response
FAIL Main resource load matched the rule with race-network-and-fetch-handler source, and the server reseponse is faster than the fetch handler assert_equals: expected "Network with GET request" but got "rrexb"
PASS Subresource load matched the rule with race-network-and-fetch-handler source, and the fetch handler response is faster than the server response
FAIL Subresource load matched the rule with race-network-and-fetch-handler source, and the server reseponse is faster than the fetch handler assert_equals: expected "Network with GET request" but got "bsmfq"
FAIL Subresource load matched the rule with race-network-and-fetch-handler source, and the server reseponse with 204 response is faster than the fetch handler assert_equals: expected 204 but got 200
PASS Subresource load matched the rule with race-network-and-fetch-handler source, and the server reseponse is faster than the fetch handler, but not found
